Struct router_rs::prelude::Router

pub struct Router { /* fields omitted */ }

Holds all of the routes for an application.

A typical use case is that when we want to move to a new route (such as after clicking on an anchor tag) we'll query our router to see if the new route matches any of our route definitions.

Then if we find a matching route we'll return it.

Implementations

impl Router

pub fn provide<T: 'static>(&mut self, provided: T)

Provide the application state data that different routes need.

impl Router

pub fn set_route_handlers(&mut self, route_handlers: Vec<Box<dyn RouteHandler>>)

Push a vector of Routes into the Router

pub fn matching_routerhandler(
&self,
incoming_route: &str
) -> Option<&Box<dyn RouteHandler>>

Return the matching RouteHandler given some incoming_route

pub fn view(&self, incoming_route: &str) -> Option<VirtualNode>

Get the first route in our routes vector view that handles this incoming_route and return the view for that route.

You'll typically call this when trying to render the correct view based on the page URL or after clicking on an anchor tag.