Module router_rs::prelude

Things that you'll usually need when working with frontend routing

Re-exports

pub use router_rs_macro::create_routes;

pub use router_rs_macro::route;

Structs

Provided
Data that was provided by the developer.

Route
A route specifies a path to match against. When a match is found a view_creator is used to return an impl View that can be used to render the appropriate content for that route.

Router
Holds all of the routes for an application.

Traits

RouteHandler
Used by router-rs-macro during code generation when turning your

RouteParam
Enables a type to be used as a route paramer

Type Definitions

ParseRouteParam
Given a param_key &str and a param_val &str, get the corresponding route parameter

ProvidedMap
A map of TypeId's to Box<Provided> (stored as Box)